requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.7.4
chromadb==0.5.3
sentence-transformers==3.0.1
anthropic==0.34.0
pandas==2.2.2
lxml==5.2.2
PyMuPDF==1.24.9
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.31
